% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{normarlity_test}
\alias{normarlity_test}
\title{Check the normaility of date
There is three components of the test
The first part is to create to histogram with desnity polt
The second part is to  create a qqplot
The part is to run the Shapiro-Wilk normality test}
\usage{
normarlity_test(data, alpha = 0.05, plotting = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{A number vector of data values}

\item{alpha=0.05}{A double numeric value, which is the significance level used for the Shapiro-Wilk test of normality.}

\item{plotting=TRUE}{A logical argument indicating whether or not to print the histogram and qq-plot.}

\item{Verbose=TRUE}{A logical argument on whether or not to give the users some feedbacks.}
}
\value{
A string that tells the users whether the data is normal is not by the Shapiro-Wilk normality test
}
\description{
Check the normaility of date
There is three components of the test
The first part is to create to histogram with desnity polt
The second part is to  create a qqplot
The part is to run the Shapiro-Wilk normality test
}
